:tada: feat: add penal code popup
