+++ packages/client/package.json
@@ -18,6 +18,7 @@
     "@headlessui/react": "^1.4.2",
     "@radix-ui/react-context-menu": "^0.1.4",
     "@radix-ui/react-dropdown-menu": "^0.1.4",
+    "@radix-ui/react-hover-card": "^0.1.3",
     "@radix-ui/react-switch": "^0.1.4",
     "@react-aria/label": "^3.2.1",
     "@react-aria/ssr": "^3.1.0",
