+++ packages/client/src/components/leo/modals/CreateTicketModal.tsx
@@ -134,6 +134,7 @@ export function CreateTicketModal({ type }: { type: RecordType }) {
 
             <FormField errorMessage={errors.violations as string} label={t("violations")}>
               <Select
+                extra={{ showPenalCodeDescriptions: true }}
                 value={values.violations}
                 name="violations"
                 onChange={handleChange}
